feature: Connect location details with the server

+++ app/src/main/java/com/nodosacademy/boldforecast/data/Day.kt
@@ -8,7 +8,7 @@ data class Day(
     @SerializedName("maxtemp_c") var maxtempC: Double? = null,
     @SerializedName("maxtemp_f") var maxtempF: Double? = null,
     @SerializedName("mintemp_c") var mintempC: Double? = null,
-    @SerializedName("mintemp_f") var mintempF: Int? = null,
+    @SerializedName("mintemp_f") var mintempF: Double? = null,
     @SerializedName("avgtemp_c") var avgtempC: Double? = null,
     @SerializedName("avgtemp_f") var avgtempF: Double? = null,
     @SerializedName("maxwind_mph") var maxwindMph: Double? = null,
